Modificación del metodo eliminar

+++ Interfaces/ctrl_ventana_superusuario.py
@@ -40,24 +40,51 @@ class MainWindow(QtWidgets.QMainWindow,Ui_MainWindow):
 			print(e)
 	def registrar_usuario(self):
 		self._main_window = ctrl_registrar_usuario.MainWindow()
+		self._main_window.set_parent(self)
 		self._main_window.show()
 	def btn_Eliminar_clicked(self):
 		ret=QMessageBox.question(self, 'Advertencia!', "¿Estas seguro de que desea eliminar este usuario?", QMessageBox.Yes | QMessageBox.No )
 		if ret == QMessageBox.Yes:
 			print('Yes clicked.')
 			admin=Admin.Admin()
-			try:
-				admin.eliminar_user(self.tableWidget.selectedItems()[0].text())
-			except IndexError:
-				ret=QMessageBox.question(self, 'Advertencia!', "TIENE QUE SELECCIONAR EL ID DEL USUARIO QUE DESEA MODIFICAR", QMessageBox.Ok )
-			except Exception as e:
-				print(e)
-			pass
+			if self.tableWidget.selectedItems()[4].text() == 'Inactivo':
+				try:
+					admin.activar_user(self.tableWidget.selectedItems()[0].text())
+					self.tableWidget.selectedItems()[4].setText('Actvo')
+				except IndexError:
+					ret = QMessageBox.question(self, 'Advertencia!',
+											   "TIENE QUE SELECCIONAR EL ID DEL USUARIO QUE DESEA MODIFICAR",
+											   QMessageBox.Ok)
+				except Exception as e:
+					print(e)
+				pass
+			else:
+				try:
+					admin.eliminar_user(self.tableWidget.selectedItems()[0].text())
+					self.tableWidget.selectedItems()[4].setText('Inactivo')
+				except IndexError:
+					ret=QMessageBox.question(self, 'Advertencia!', "TIENE QUE SELECCIONAR EL ID DEL USUARIO QUE DESEA MODIFICAR", QMessageBox.Ok )
+				except Exception as e:
+					print(e)
+				pass
 		else:
 			print('No clicked.')
 			pass
 	def limpiar_tabla(self):
 		self.tableWidget.clearContents()
+	def modificar_fila(self, id, nombre, email, role):
+		row = self.tableWidget.currentRow()
+		self.tableWidget.setItem(row, 0, QtWidgets.QTableWidgetItem(id))
+		self.tableWidget.setItem(row, 1, QtWidgets.QTableWidgetItem(nombre))
+		self.tableWidget.setItem(row, 2, QtWidgets.QTableWidgetItem(email))
+		self.tableWidget.setItem(row, 3, QtWidgets.QTableWidgetItem(role))
+	def add_fila(self, id, nombre, email, role):
+		rowPosition = self.tableWidget.rowCount()
+		self.tableWidget.insertRow(rowPosition)
+		self.tableWidget.setItem(rowPosition, 0, QtWidgets.QTableWidgetItem(id))
+		self.tableWidget.setItem(rowPosition, 1, QtWidgets.QTableWidgetItem(nombre))
+		self.tableWidget.setItem(rowPosition, 2, QtWidgets.QTableWidgetItem(email))
+		self.tableWidget.setItem(rowPosition, 3, QtWidgets.QTableWidgetItem(role))
 	def load_usuarios(self):
 		try:
 			admin=Admin.Admin()
@@ -70,7 +97,16 @@ class MainWindow(QtWidgets.QMainWindow,Ui_MainWindow):
 					self.tableWidget.setItem(rowPosition,0,QtWidgets.QTableWidgetItem(results['id'][i]))
 					self.tableWidget.setItem(rowPosition,1,QtWidgets.QTableWidgetItem(results['username'][i]))
 					self.tableWidget.setItem(rowPosition,2,QtWidgets.QTableWidgetItem(results['email'][i]))
-					self.tableWidget.setItem(rowPosition,3,QtWidgets.QTableWidgetItem(results['role'][i]))
+					if results['role'][i] == str(0):
+						self.tableWidget.setItem(rowPosition, 3, QtWidgets.QTableWidgetItem('user'))
+					else:
+						self.tableWidget.setItem(rowPosition, 3, QtWidgets.QTableWidgetItem('admin'))
+					#self.tableWidget.setItem(rowPosition,3,QtWidgets.QTableWidgetItem(results['role'][i]))
+					if results['Actividad'][i] == str(0):
+						self.tableWidget.setItem(rowPosition, 4, QtWidgets.QTableWidgetItem('Inactivo'))
+					else:
+						self.tableWidget.setItem(rowPosition, 4, QtWidgets.QTableWidgetItem('Activo'))
+
 			except Exception as n:
 				print (n)
 				print ('cargar datos tabla')
